+++ base/napi/binary/include/stdlib/math/base/napi/binary/uu_u.h
@@ -1,7 +1,7 @@
 /**
 * @license Apache-2.0
 *
-* Copyright (c) 2020 The Stdlib Authors.
+* Copyright (c) 2025 The Stdlib Authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -16,38 +16,38 @@
 * limitations under the License.
 */
 
-#ifndef STDLIB_MATH_BASE_NAPI_BINARY_II_D_H
-#define STDLIB_MATH_BASE_NAPI_BINARY_II_D_H
+#ifndef STDLIB_MATH_BASE_NAPI_BINARY_UU_U_H
+#define STDLIB_MATH_BASE_NAPI_BINARY_UU_U_H
 
 #include <node_api.h>
 #include <assert.h>
 #include <stdint.h>
 
 /**
-* Macro for registering a Node-API module exporting an interface invoking a binary function accepting signed 32-bit integers and returning a double-precision floating-point number.
+* Macro for registering a Node-API module exporting an interface invoking a binary function accepting and returning unsigned 32-bit integers.
 *
 * @param fcn   binary function
 *
 * @example
 * #include <stdint.h>
 *
-* static double add( const int32_t x, const int32_t y ) {
+* static uint32_t add( const uint32_t x, const uint32_t y ) {
 *     return x + y;
 * }
 *
 * // ...
 *
 * // Register a Node-API module:
-* STDLIB_MATH_BASE_NAPI_MODULE_II_D( add );
+* STDLIB_MATH_BASE_NAPI_MODULE_UU_U( add );
 */
-#define STDLIB_MATH_BASE_NAPI_MODULE_II_D( fcn )                               \
-	static napi_value stdlib_math_base_napi_ii_d_wrapper(                      \
+#define STDLIB_MATH_BASE_NAPI_MODULE_UU_U( fcn )                               \
+	static napi_value stdlib_math_base_napi_uu_u_wrapper(                      \
 		napi_env env,                                                          \
 		napi_callback_info info                                                \
 	) {                                                                        \
-		return stdlib_math_base_napi_ii_d( env, info, fcn );                   \
+		return stdlib_math_base_napi_uu_u( env, info, fcn );                   \
 	};                                                                         \
-	static napi_value stdlib_math_base_napi_ii_d_init(                         \
+	static napi_value stdlib_math_base_napi_uu_u_init(                         \
 		napi_env env,                                                          \
 		napi_value exports                                                     \
 	) {                                                                        \
@@ -56,14 +56,14 @@
 			env,                                                               \
 			"exports",                                                         \
 			NAPI_AUTO_LENGTH,                                                  \
-			stdlib_math_base_napi_ii_d_wrapper,                                \
+			stdlib_math_base_napi_uu_u_wrapper,                                \
 			NULL,                                                              \
 			&f                                                                 \
 		);                                                                     \
 		assert( status == napi_ok );                                           \
 		return f;                                                              \
 	};                                                                         \
-	NAPI_MODULE( NODE_GYP_MODULE_NAME, stdlib_math_base_napi_ii_d_init )
+	NAPI_MODULE( NODE_GYP_MODULE_NAME, stdlib_math_base_napi_uu_u_init )
 
 /*
 * If C++, prevent name mangling so that the compiler emits a binary file having undecorated names, thus mirroring the behavior of a C compiler.
@@ -73,12 +73,12 @@ extern "C" {
 #endif
 
 /**
-* Invokes a binary function accepting signed 32-bit integers and returning a double-precision floating-point number.
+* Invokes a binary function accepting and returning unsigned 32-bit integers.
 */
-napi_value stdlib_math_base_napi_ii_d( napi_env env, napi_callback_info info, double (*fcn)( int32_t, int32_t ) );
+napi_value stdlib_math_base_napi_uu_u( napi_env env, napi_callback_info info, uint32_t (*fcn)( uint32_t, uint32_t ) );
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif // !STDLIB_MATH_BASE_NAPI_BINARY_II_D_H
+#endif // !STDLIB_MATH_BASE_NAPI_BINARY_UU_U_H